requirements.txt
behave==1.2.5
Django==1.10.1
django-behave==0.1.5
factory-boy==2.7.0
fake-factory==0.7.2
parse==1.6.6
parse-type==0.3.4
python-dateutil==2.5.3
selenium==2.53.6
six==1.10.0
splinter==0.7.4
